First draft of observation edition functionality
Created Observation Component, which listens to the store and takes care
of sending the props to the other components. It also holds the current
observation in edition in the state. When all edition is done, it should
be the one updating the store.

Changed the ObservationView in a functional component.

Created CategorySelect as a functional component. It receives a callback
function from Observation, which updates the Category in the store.

Created stubs for location_edit and details_edit